[ci] refactor state machine into a base class
